+++ python/python/tests/test_transaction.py
@@ -8,7 +8,13 @@ import pytest
 from glide import RequestError
 from glide.async_commands.bitmap import BitmapIndexType, OffsetOptions
 from glide.async_commands.command_args import Limit, ListDirection, OrderBy
-from glide.async_commands.core import InsertPosition, StreamAddOptions, TrimByMinId
+from glide.async_commands.core import (
+    ExpiryGetEx,
+    ExpiryTypeGetEx,
+    InsertPosition,
+    StreamAddOptions,
+    TrimByMinId,
+)
 from glide.async_commands.sorted_set import (
     AggregationType,
     GeoSearchByRadius,
@@ -58,6 +64,7 @@ async def transaction_test(
     key18 = "{{{}}}:{}".format(keyslot, get_random_string(3))  # sort
     key19 = "{{{}}}:{}".format(keyslot, get_random_string(3))  # bitmap
     key20 = "{{{}}}:{}".format(keyslot, get_random_string(3))  # bitmap
+    key22 = "{{{}}}:{}".format(keyslot, get_random_string(3))  # getex
 
     value = datetime.now(timezone.utc).strftime("%m/%d/%Y, %H:%M:%S")
     value2 = get_random_string(5)
@@ -419,6 +426,15 @@ async def transaction_test(
         alpha=True,
     )
     args.append(4)
+
+    min_version = "6.2.0"
+    if not await check_if_server_version_lt(redis_client, min_version):
+        transaction.set(key22, "value")
+        args.append(OK)
+        transaction.getex(key22)
+        args.append("value")
+        transaction.getex(key22, ExpiryGetEx(ExpiryTypeGetEx.SEC, 1))
+        args.append("value")
 
     min_version = "7.0.0"
     if not await check_if_server_version_lt(redis_client, min_version):